# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "tripleS Home Automation API",
        "version": "1.0.0",
        "contact": {}
    },
    "basePath": "/v1/",
    "paths": {
...
        "/auth/forgetPassword/getCode": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Get code",
                "operationId": "AuthController.GetCode",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The body content",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetCodeForm"
                        }
                    }
                ],
                "responses": {
                    "100": {
                        "description": "wrong format"
                    },
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/auth/forgetPassword/verifyCode": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Verify Code",
                "operationId": "AuthController.VerifyCode",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The object content",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.VerifyCodeForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "403": {
                        "description": "code not found"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "User Login",
                "operationId": "AuthController.UserLogin",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "uid",
                        "description": "apikey",
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "pwd",
                        "description": "apikey",
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": "wrong format"
                    },
                    "200": {
                        "schema": {
                            "$ref": "#/definitions/model.LoginRsp"
                        }
                    },
                    "410": {
                        "description": "wrong password"
                    }
                }
            }
        },
...
        "/auth/tokenVerify": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Token verfiy",
                "operationId": "AuthController.TokenVerify",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "description": "The token id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": "wrong format"
                    },
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "token is expired"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.GetCodeForm": {
            "title": "GetCodeForm",
            "type": "object",
            "properties": {
                "uid": {
                    "type": "string"
                }
            }
        },
...
        "model.LoginRsp": {
            "title": "LoginRsp",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int64"
                },
                "data": {
                    "$ref": "#/definitions/model.UserInfo"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "model.UserInfo": {
            "title": "UserInfo",
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "model.VerifyCodeForm": {
            "title": "VerifyCodeForm",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        }
    },
...
}